New helper visitor to fold types.
This is somewhere between TypeQueryVisitor and a general visitor -- the idea is
to handle the recursing down a tree of types to get some piece of information
out, and you provide the combining function to aggregate that information from
any two different types, and a "zero" value to start with (and a default value
for a type).

This can be used to, say, pull out a single type variable id, or all the type
variables, or whatever.