Actually erase types when we have no substitutions; substitute all args.
Before, we were not correctly detecting splat-args on the input, and treating
them as a single argument.  This resulted in giving incorrect argument-count
error messages. Fixes that problem by detecting splat-args and not making any
assertions about the number of substitutions in that case.

Also, we were not actually substituting for all the arguments, and now we are.